pos_system/requirements.txt
# Pinned versions — update only after reviewing changelogs for security fixes
bcrypt==5.0.0
pillow==12.1.1
python-docx==1.1.2
# Not yet installed — install before enabling barcode/export features
# reportlab
# openpyxl
# python-barcode
